Hacking makefiles.  Addresses #782.  Also fixes a valgrind bug in one of the recovery tests.
git-svn-id: file:///svn/tokudb@3814 c7de825b-a66e-492c-adef-691d508d4ae1

+++ src/tests/test_logmax.c
@@ -50,7 +50,7 @@ void test_logmax (int logmax) {
 	r=env->get_lg_max(env, &lmax);
 	assert(r==0);
 	if (logmax>0) {
-	    assert(lmax==logmax);
+	    assert(lmax==(u_int32_t)logmax);
 	} else {
 	    assert(lmax>0);
